feat: extract database and fix coroutines
